Add files via upload
The EipTagListFactory class is a generic wrapper to read all tags and the templates of all structured tags of an Ethernet/IP PLC. The creation of the EipTagList (tree list)  containing all controller and program tags of the PLC is available via an optional method. Please bear in mind: The wrapper works fine for CompactLogix and ControlLogix with an Ethernet connection to a PLC onboard Ethernet adapter. "Slot-routing" is not supported as I didn't find a documentation how to create the CIP path message frame. The wrapper is creating his own TcpClient connection to the PLC to get the needed information. The constructor creates the communication session, reads the information and closes the communication session. All needed information is available via the public methods and members.