Implemented Tag Browsing
Implemented Managed Tag Browsing Capabilities. All this is implemented in the "libPlcTag.browse.TagBrowser" class. This class uploads and parses all Controller-Tags, Udts and Program-Tags.

To visualize all this, there is now also an "Tag Browser" application under "Examples" that shows how this works

This is currently only teste with an "Softlogix 5800 V19" controller, so it should work with Compact and ControlLogix Plcs in general. Unfortunately i do not have access to any MicroLogix Cpu, so i could not test this with this type of controller. Since the Architecture of Micrologix is vastly different to Compact/ControlLogix, I do not expect this to work with Micrologix.